Second commit: started with the phone design.
Overall: created a structure for the only view; the Home.js page.
Created all the files needed for the components About, Skills, etc.
Created two other components: DoubleBorder for in between components,
and NextArrow for an indication of going to the next page. As every section will
hopefully be 100vh, so the idea is that you cannot see what's below.

Navigation: created a nav with a double border and a white background, has
also a text which serves as an indication of where you are, and a burger menu.
The nav content is also there but the icon doesn't work yet.

Header: has just some random text, have no clue just yet what to put there. Has
the Next Arrow component, which will go to the About page.

Footer: has only text but serves as a link to the contact component overlay.

:)