Portfolio is now at elmaharmsen.nl;
Changed the links for the about me images and the link to my
previous portfolio.
Added a js media query and with this tried to make the scrolling to work
with different nav heights for phone and laptop. (Shit doesn't work)
Changed .App at phone from display flex to grid.